Create new HumanLayer class for global costmap
Create a 2D costmap layer plugin with a simpler cost function to
represent humans in the global costmap. This inherits from the
SocialLayer class and overrides the cost function to a simple lethal
circle, an inscribed cost circle and an exponential decay. This allows
us to better control the spacing the global planner gives to people
compared to the ProxemicLayer, which uses a gaussian cost function.

ROS-781